loan-management-system/src/requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-phonenumber-field==7.3.0
django-rosetta==0.10.0
gunicorn==21.2.0
idna==3.6
jalali_core==1.0.0
jdatetime==5.0.0
packaging==23.2
phonenumbers==8.13.27
polib==1.2.0
psycopg2==2.9.9
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
